Add yaml validation to code editor
